refactor: fold DiscreteTimeSystem into HiddenMarkovModel
Remove the intermediate discrete-time base and keep HMM-specific helpers
(time_step check, one-hot) on HiddenMarkovModel
